requirements.txt
sqlalchemy==2.0.21
argon2-cffi==23.1.0
click==8.1.7
python-dotenv==1.0.0
sentry-sdk==1.32.0
pytest==7.4.2
pytest-cov==4.1.0